nit: hide out-of-scope function
