--- utils/functions/getCurrentShowCachedTranslations.ts
@@ -1,6 +1,0 @@
-import getAllCachedTranslations from "./getAllCachedTranslations"
-
-export default async function getCurrentShowCachedTranslations() {
-    const cache = await getAllCachedTranslations()
-    return cache?.[window.currentShowId]?.sentences || {}
-}
